+++ src/app.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { PlayersModule } from './players/players.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Player } from './players/entities/players.entity';
-import { Region } from './entities/app.entity';
+import { Player } from './players/entities/player.entity';
+import { Region } from './entities/region.entity';
 import { Position } from './players/entities/position.entity';
 import { PitchStyle } from './players/entities/pitch.entity';
 import { BattingSide } from './players/entities/batting.entity';
